revert: nullable Quote (#799)

+++ src/s-z/Stc/Stc.Models.cs
@@ -3,5 +3,5 @@ namespace Skender.Stock.Indicators;
 [Serializable]
 public class StcResult : ResultBase
 {
-    public decimal? Stc { get; set; }
+    public double? Stc { get; set; }
 }
